+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to the "robotcode" extension will be documented in this file
 - Added some more configuration options for log and debug messages when running tests in the debug console
 - debug console now shows source and line number from log messages
 - use of debugpy from vscode Python extension, no separate installation of debugpy required
+- implement test tags in test controller
 
 ##  0.2.7